# swagger.yaml
openapi: 3.0.0
info:
  title: OSS Project API
  description: API documentation for the OSS Project, including authentication, project management, document, and page management.
  version: '1.0'
servers:
  - url: https://oss-project-api.onrender.com/api/v1
    description: Production server
  - url: http://localhost:3000/api/v1
    description: Development server
paths:
...
  /documents:
    get:
      tags:
        - Document
      summary: Get all documents
      operationId: getAllDocuments
      responses:
        '200':
          description: List of all documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'

    post:
      tags:
        - Document
      summary: Create a new document
      operationId: createDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDocumentRequest'
      responses:
        '201':
          description: Document created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
  /documents/{documentId}:
    get:
      tags:
        - Document
      summary: Get document by ID
      operationId: getDocumentById
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Document details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'

    patch:
      tags:
        - Document
      summary: Update document by ID
      operationId: updateDocument
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDocumentRequest'
      responses:
        '200':
          description: Document updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'

    delete:
      tags:
        - Document
      summary: Delete document by ID
      operationId: deleteDocument
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Document deleted successfully
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Document:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        project:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateDocumentRequest:
      type: object
      required:
        - title
        - content
        - project
      properties:
        title:
          type: string
        content:
          type: string
        project:
          type: string
    UpdateDocumentRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        project:
          type: string
security:
  - bearerAuth: []
...